copy PRU_SUART_Emulation from lms2012
This adds the assembly code from a dump of the PRU_SUART_Emulation.p
program from the lms2012 source code. This creates nearly identical
firmware with the exception of a slight difference in the encoding
of the NOT instruction (bit 24 is cleared instead of set indicating
that Op2 is a register instead of immediate, however this instruction
is unary and doesn't use Op2) and some extra instructions for the C
main() wrapper is added.

Tooling is updated to v2.3.3.